* directory.m (java_ensure_selector): Use unsigned for `pos' argument to addstr to avoid potential confusion with %u.
git-svn-id: svn://svn.savannah.nongnu.org/swarm/trunk@4080 84d386b0-d280-4341-8980-c4cb7e1eb7a5

+++ swarm/src/defobj/directory.m
@@ -455,7 +455,7 @@ java_ensure_selector (JNIEnv *env, jobject jsel)
             return (pos + alignment) & ~mask;
         }
 
-      void addstr (char type, BOOL regFlag, size_t pos)
+      void addstr (char type, BOOL regFlag, unsigned pos)
         {
           *p++ = type;
           if (regFlag)
